agent/requirements.txt
fastapi
uvicorn
langchain
python-dotenv
requests
pydantic
weaviate-client
